unify checks in a single file
